requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   fiona
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
bqplot==0.12.44
    # via geemap
branca==0.8.1
    # via
    #   folium
    #   ipyleaflet
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.1.31
    # via
    #   fiona
    #   pyogrio
    #   pyproj
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.1
    # via requests
click==8.1.8
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   geocoder
    #   streamlit
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
colorama==0.4.6
    # via radon
colour==0.1.5
    # via geemap
comm==0.2.2
    # via
    #   ipykernel
    #   ipywidgets
contourpy==1.3.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via
    #   ipython
    #   ratelim
distlib==0.3.9
    # via virtualenv
earthengine-api==1.5.9
    # via
    #   eerepr
    #   geemap
eerepr==0.1.1
    # via geemap
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via virtualenv
fiona==1.10.1
    # via ciut-tablero (pyproject.toml)
folium==0.19.5
    # via geemap
fonttools==4.57.0
    # via matplotlib
future==1.0.0
    # via geocoder
geemap==0.35.3
    # via ciut-tablero (pyproject.toml)
geocoder==1.38.1
    # via geemap
geopandas==1.0.1
    # via ciut-tablero (pyproject.toml)
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-api-core==2.24.2
    # via
    #   google-api-python-client
    #   google-cloud-core
    #   google-cloud-storage
google-api-python-client==2.167.0
    # via earthengine-api
google-auth==2.38.0
    # via
    #   earthengine-api
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-cloud-core
    #   google-cloud-storage
google-auth-httplib2==0.2.0
    # via
    #   earthengine-api
    #   google-api-python-client
google-cloud-core==2.4.3
    # via google-cloud-storage
google-cloud-storage==3.1.0
    # via earthengine-api
google-crc32c==1.7.1
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.7.2
    # via google-cloud-storage
googleapis-common-protos==1.70.0
    # via google-api-core
httplib2==0.22.0
    # via
    #   earthengine-api
    #   google-api-python-client
    #   google-auth-httplib2
identify==2.6.9
    # via pre-commit
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
ipyevents==2.0.2
    # via geemap
ipyfilechooser==0.6.0
    # via geemap
ipykernel==6.29.5
    # via ciut-tablero (pyproject.toml)
ipyleaflet==0.19.2
    # via geemap
ipython==9.1.0
    # via
    #   ipykernel
    #   ipywidgets
ipython-pygments-lexers==1.1.1
    # via ipython
ipytree==0.2.2
    # via geemap
ipywidgets==8.1.6
    # via
    #   bqplot
    #   ipyevents
    #   ipyfilechooser
    #   ipyleaflet
    #   ipytree
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   branca
    #   folium
    #   pydeck
jsonschema==4.23.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.7.2
    # via
    #   ipykernel
    #   jupyter-client
jupyter-leaflet==0.19.2
    # via ipyleaflet
jupyterlab-widgets==3.0.14
    # via ipywidgets
kiwisolver==1.4.8
    # via matplotlib
mando==0.7.1
    # via radon
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.1
    # via
    #   geemap
    #   seaborn
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
narwhals==1.34.1
    # via
    #   altair
    #   plotly
nest-asyncio==1.6.0
    # via ipykernel
nodeenv==1.9.1
    # via pre-commit
numpy==2.2.4
    # via
    #   bqplot
    #   contourpy
    #   folium
    #   geemap
    #   geopandas
    #   matplotlib
    #   pandas
    #   pydeck
    #   pyogrio
    #   seaborn
    #   shapely
    #   streamlit
packaging==24.2
    # via
    #   altair
    #   geopandas
    #   ipykernel
    #   matplotlib
    #   plotly
    #   pyogrio
    #   pytest
    #   streamlit
pandas==2.2.3
    # via
    #   bqplot
    #   geemap
    #   geopandas
    #   seaborn
    #   streamlit
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via
    #   matplotlib
    #   streamlit
pip==25.0.1
    # via ciut-tablero (pyproject.toml)
platformdirs==4.3.7
    # via
    #   jupyter-core
    #   virtualenv
plotly==6.0.1
    # via geemap
pluggy==1.5.0
    # via pytest
pre-commit==4.2.0
    # via ciut-tablero (pyproject.toml)
prompt-toolkit==3.0.50
    # via ipython
proto-plus==1.26.1
    # via google-api-core
protobuf==5.29.4
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==19.0.1
    # via streamlit
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pydeck==0.9.1
    # via streamlit
pygments==2.19.1
    # via
    #   ipython
    #   ipython-pygments-lexers
pyogrio==0.10.0
    # via geopandas
pyparsing==3.2.3
    # via
    #   httplib2
    #   matplotlib
pyperclip==1.9.0
    # via geemap
pyproj==3.7.1
    # via geopandas
pyshp==2.3.1
    # via geemap
pytest==8.3.5
    # via ciut-tablero (pyproject.toml)
python-box==7.3.2
    # via geemap
python-dateutil==2.9.0.post0
    # via
    #   jupyter-client
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pre-commit
pyzmq==26.4.0
    # via
    #   ipykernel
    #   jupyter-client
radon==6.0.1
    # via ciut-tablero (pyproject.toml)
ratelim==0.1.6
    # via geocoder
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.3
    # via
    #   earthengine-api
    #   folium
    #   geocoder
    #   google-api-core
    #   google-cloud-storage
    #   streamlit
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9
    # via google-auth
scooby==0.10.0
    # via geemap
seaborn==0.13.2
    # via ciut-tablero (pyproject.toml)
setuptools==78.1.0
    # via ciut-tablero (pyproject.toml)
shapely==2.1.0
    # via geopandas
six==1.17.0
    # via
    #   geocoder
    #   mando
    #   python-dateutil
smmap==5.0.2
    # via gitdb
stack-data==0.6.3
    # via ipython
streamlit==1.44.1
    # via ciut-tablero (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.4.2
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
traitlets==5.14.3
    # via
    #   bqplot
    #   comm
    #   ipykernel
    #   ipython
    #   ipywidgets
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
    #   traittypes
traittypes==0.2.1
    # via
    #   bqplot
    #   ipyleaflet
typing-extensions==4.13.2
    # via
    #   altair
    #   referencing
    #   streamlit
tzdata==2025.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.4.0
    # via requests
virtualenv==20.30.0
    # via pre-commit
vulture==2.14
    # via ciut-tablero (pyproject.toml)
watchdog==6.0.0
    # via streamlit
wcwidth==0.2.13
    # via prompt-toolkit
widgetsnbextension==4.0.14
    # via ipywidgets
xyzservices==2025.1.0
    # via
    #   folium
    #   ipyleaflet
